refactor: Deleted PagePaginateLite and added into PagePaginate
