+++ src/Wrappers/PagePaginate.cs
@@ -32,4 +32,29 @@ public static class PagePaginate
 
         return task;
     }
+    
+    public static Task<Result<Page<T>>> Paginate<T>(
+        this IQueryable<T> query,
+        int page,
+        int limit)
+    {
+        var length = query.Count();
+
+        if (length == 0)
+        {
+            return Task.FromResult(Result.Ok(Page<T>.Set(page, [])));
+        }
+
+        var list = query
+            .Skip(page * limit)
+            .Take(limit)
+            .ToList();
+
+        var task = new Task<Result<Page<T>>>(() => Result.Ok(
+            Page<T>.Set(length, list)));
+
+        task.RunSynchronously();
+
+        return task;
+    }
 }
